bigfile/py: PyFrameObject->{f_code,f_locals,f_localsplus} can no longer be directly accessible on py3.11
    bigfile/_bigfile.c: In function ‘pybigfile_loadblk’:
    bigfile/_bigfile.c:736:35: error: invalid use of incomplete typedef ‘PyFrameObject’ {aka ‘struct _frame’}
      736 |                     fastlocals = f->f_localsplus;
          |                                   ^~
    bigfile/_bigfile.c:737:31: error: invalid use of incomplete typedef ‘PyFrameObject’ {aka ‘struct _frame’}
      737 |                     for (j = f->f_code->co_nlocals; j >= 0; --j) {
          |                               ^~
    bigfile/_bigfile.c:746:26: error: invalid use of incomplete typedef ‘PyFrameObject’ {aka ‘struct _frame’}
      746 |                     if (f->f_locals != NULL) {
          |                          ^~

https://docs.python.org/3/whatsnew/3.11.html says to use
PyFrame_GetCode() instead of ->f_code, PyFrame_GetLocals() instead of
->f_locals, and that there is no public API for ->f_localsplus.

For ->f_code we do as suggested.

For ->f_localsplus we do direct access via frame->f_frame->localsplus.
This is internal API available only under Py_BUILD_CORE but Cython and
other projects use this everywhere:

    https://github.com/cython/cython/blob/066adcb3/Cython/Utility/Exceptions.c#L849-L857
    https://sources.debian.org/src/numba/0.61.2+dfsg-1/numba/_dispatcher.cpp/?hl=35#L35
    https://sources.debian.org/src/systemtap/5.1-4.1/python/HelperSDT/_HelperSDT.c/?hl=154#L154
    ...

so we are ok to use this private bits as well.

For ->f_locals we do not use PyFrame_GetLocals() because we want to
check whether ->f_locals is NULL or not NULL while PyFrame_GetLocals
creates new dict and fills it when ->f_locals was NULL. This way we,
similarly to f_localsplus, use direct PyFrameObject->_PyInterpreterFrame->f_locals
to do the access.