WCFS: Allow file truncation at transaction boundary
Previously, WCFS code enforced the invariant that files only grow or
remain the same size and never shrink [1]. This patch relaxes that
invariant by allowing files to shrink, but only at transaction
boundaries. Within a single transaction, files still only grow or
persist in size.

This change is essential to support a forthcoming .discard_data method
for ZBigFile. To ensure safety with this relaxed invariant, the patch:

- zeroes truncated regions during resynchronization on connections where
  truncation occurred
- pins blocks to their previous versions on concurrent connections to
  those that discarded data at the transaction boundary

It remains safe to mmap regions beyond the file size with zeros, as file
shrinking is restricted to transaction boundaries.

[1] https://lab.nexedi.com/nexedi/wendelin.core/-/blob/e8a00ac0/wcfs/client/wcfs.cpp#L68-72