wcfs: tests: Increase each test timeout from 10s to 15s
Each wcfs test is run under timeout to detect e.g. that something is
stuck and to unmount the filesystem forcibly on such case. That overall
timeout is much less compared to regular 30s pinkill timeout. And is
much high compared to 3s pinkill timeout used in faulty-protection tests.

The faulty protection tests usually need to wait for a kill to happen
for 2·pinkill timeout to reliably detect that event in the presence of
surrounding OS load and that was working quite ok so far because the
actual kill was usually happening in around 1·pinkill since start of
the waiting. However in the next patch we will need to test whether wcfs
does _not_ kill an innocent client, similarly waiting for that 2·pinkill
time, but here it will be full 2·pinkill sleep without trimming because
for negative condition (does _not_ kill) we need to predictably wait
much longer compared to pinkill time. With that 6s just for the sleep,
and test setup and other overhead things start to trigger overall
timeout quite frequently.

-> Increase that overall timeout from 10s to 15s to cover that "need to
   wait longer" situation while still maintaining invariant for the timeout
   to stay much less of regular pinkill time and much more of faultyprot
   tests pinkill time.

/reviewed-by @levin.zimmermann
/reviewed-on https://lab.nexedi.com/nexedi/wendelin.core/-/merge_requests/33